Layout consistency change and small debugs

+++ src/frontend/AdminEventAction/DeleteEvent.js
@@ -6,7 +6,10 @@ const DeleteEvent = () => {
     const submitDelete = async (e) => {
         e.preventDefault();
 
-        let eventId = document.getElementById("eventId").value;
+        let eventId = parseInt(document.getElementById("eventId").value);
+        
+        if (!eventId)
+            return toast.error("Event ID is invalid.");
 
         await fetch(`${process.env.REACT_APP_SERVER_URL}/event/delete`, {
             method: "DELETE",
@@ -30,10 +33,12 @@ const DeleteEvent = () => {
     // displayed when "Delete location" button is clicked
     return(
         <form>
-        <h4>Delete event</h4><hr/>
-        <label htmlFor="event"></label>
-        <input type="text" id="eventId" name="eventId" placeholder='Enter an event ID'/>
-        <button className="btn btn-success" onClick={(e) => {submitDelete(e)}}>Delete</button>
+
+            <h4>Delete event</h4><hr/>
+
+            <label htmlFor="event"></label>
+            <input type="text" id="eventId" name="eventId" placeholder='Enter an event ID'/>
+            <button className="btn btn-success" onClick={(e) => {submitDelete(e)}}>Delete</button>
   
         </form>
     )
